Move raw FFI interface to gettext-sys crate
